overriding default node options
we are providing custom node options, which are basically
the same as the default ones except for the fact that we
remove the default node name (the one passed via command line
arguments). This way, the plugin node name is actually initialized
with the one passed as an argument to the node constructor.